requirements.txt
requests
beautifulsoup4
langdetect
mistralai
python-dotenv
